+++ checkpoint/CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Context manager helper functions for Checkpointer and CheckpointManager.
 - Protobuf metadata saved by PyTreeCheckpointHandler.
 
+### Changed
+- Allow calling `create_coordinator_server_and_context` without an initialized
+JAX coordinator server and create TS metadata for numpy arrays on a single process.
+
 ## [0.2.6] - 2022-06-28
 
 ### Added